Написал Import и Export уровней и денег

+++ Assets/Scripts/SaveLoadSystem/ProgressWallet.cs
@@ -2,6 +2,6 @@ namespace Saves
 {
     public class ProgressWallet
     {
-        public byte MoneyCount;
+        public int MoneyCount;
     }
 }
